+++ apps/api/views/task_views.py
@@ -1,4 +1,5 @@
 import logging
+import sys
 
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.decorators import api_view, permission_classes
@@ -10,12 +11,7 @@ from celery.result import AsyncResult
 from drf_yasg import openapi
 from drf_yasg.utils import swagger_auto_schema
 
-from django.core.exceptions import ObjectDoesNotExist
-from django.db import IntegrityError
-
 from ..serializers import TaskSerializer
-
-from ..services.rate_service import get_rate_by_user_service, create_rate_by_user_service
 
 logger = logging.getLogger('django')
 
@@ -49,5 +45,5 @@ def read_status_task_view(request, task_id):
     task_result = AsyncResult(task_id)
     serializer = TaskSerializer(task_result)
     message = 'Запрос выполнен успешно'
-    logger.info(f'{__name__} - {message}')
+    logger.info(f'{__name__}.{sys._getframe().f_code.co_name} - {message}')
     return Response(serializer.data, status=status.HTTP_200_OK)